+++ bukkit/src/main/java/de/maxhenkel/voicechat/integration/commodore/CommodoreCommands.java
@@ -15,8 +15,10 @@ public class CommodoreCommands {
                 .then(LiteralArgumentBuilder.literal("help"))
                 .then(LiteralArgumentBuilder.literal("test").then(RequiredArgumentBuilder.argument("target", playerArgument())))
                 .then(LiteralArgumentBuilder.literal("invite").then(RequiredArgumentBuilder.argument("target", playerArgument())))
-                .then(LiteralArgumentBuilder.literal("join").then(RequiredArgumentBuilder.argument("group", uuidArgument())))
-                .then(LiteralArgumentBuilder.literal("join").then(RequiredArgumentBuilder.argument("group", uuidArgument()).then(RequiredArgumentBuilder.argument("password", StringArgumentType.string()))))
+                .then(LiteralArgumentBuilder.literal("join").then(RequiredArgumentBuilder.argument("group_id", uuidArgument())))
+                .then(LiteralArgumentBuilder.literal("join").then(RequiredArgumentBuilder.argument("group_id", uuidArgument()).then(RequiredArgumentBuilder.argument("password", StringArgumentType.string()))))
+                .then(LiteralArgumentBuilder.literal("join").then(RequiredArgumentBuilder.argument("group_name", StringArgumentType.string()).suggests(GroupNameSuggestionProvider.INSTANCE)))
+                .then(LiteralArgumentBuilder.literal("join").then(RequiredArgumentBuilder.argument("group_name", StringArgumentType.string()).suggests(GroupNameSuggestionProvider.INSTANCE).then(RequiredArgumentBuilder.argument("password", StringArgumentType.string()))))
                 .then(LiteralArgumentBuilder.literal("leave"));
 
         commodore.register(literalBuilder);